pulsemeter/requirements.txt
pyinstaller
pillow
psutil
sounddevice
numpy
pystray
pyyaml
